debug dtype

+++ data_loader/dataLoader.py
@@ -7,10 +7,10 @@ class Dataset:
         self.wav_folder = pickle.load(open('./data_loader/BZNSYP.pkl', "rb"))
     def get_item(self, index):
         item = ['p001']
-        tmp = np.arange(256)
+        tmp = np.arange(256, dtype='float64')
         tmp[0] = 1
         item.append(tmp)
-        item.append(np.resize(self.wav_folder[index][1], (256, 80)))
+        item.append(np.resize(self.wav_folder[index][1], (256, 80)).astype('float64'))
         return item
     def len(self):
         return len(self.wav_folder)
